looprpc: add get and set liquidity config endpoints
This commit adds the ability to update our liquidity config at runtime.
The decision is made to use a single LiquidityRule message with a type
enum to represent our Rule interface rather than a oneof, which can be
difficult to handle. As we add rules, the rpc will need to validate that
the correct fields are set for each enum type.