liquidity: add swap selection logic
Add functions to select channels and amounts that we should swap based
on our current imbalance and the min/max swap parameters. This version
only focuses on selecting for swaps that will be paid with a single
HTLC, as it is the more simple approach, and is available for both
swap types. Selection of channels to allow for multi-out is deferred to
a followup, with a TODO added to make note of this.