liquidity: add option to examine liquidity on a per-peer basis
This commit adds an option to the liquidity manager which allows
aggregation of channel balances on a per-peer level. This allows
liquidity rules like "I want at least 70% inbound with peer x".
To avoid the need to create a complex rules engine, this setting
is made exclusive of the per-channel rules that we previously had
in place. This is also practical because we would need to account
for pending open channels where we do not know the short chan ID
yet, and pending close channels where we would need to do multiple
lookups to match them to short channel IDs.